fix(server-capailities): override init method in ServerController
