+++ lib/api/controllers/baseController.js
@@ -62,10 +62,6 @@ class NativeController extends BaseController {
 
     this.ask = global.kuzzle.ask.bind(global.kuzzle);
     this.pipe = global.kuzzle.pipe.bind(global.kuzzle);
-    this._info = {
-      elasticsearch: {},
-      redis: {}
-    };
     this.__actions = new Set(actions);
   }
 
@@ -76,13 +72,7 @@ class NativeController extends BaseController {
    *
    * @returns {Promise}
    */
-  async init () {
-    const [storageInfo, cacheInfo] = await Promise.all([
-      global.kuzzle.ask('core:storage:public:info:get'),
-      global.kuzzle.ask('core:cache:public:info:get'),
-    ]);
-    this._info.elasticsearch = storageInfo;
-    this._info.redis = cacheInfo;
+  init () {
     return Bluebird.resolve();
   }
